Model directives (#56)
* checkpoint.  multiple collision checkers implement all of the tests.  but the number of models is getting a bit unwieldy.  going to take the new model directives for a spin

* Starts transition to model directives

Also updates pose.ipynb to use camera_box.sdf (and stop using scipy learning #47).  Still need to fix the uses in the psets, and then remove it from requirements.txt.

Starting to build up the local yaml descriptions of the scenes we've been using.

* Fully functional model directives.  Cleaned up lots of code throughout.

Also got the slider with callbacks working with the floating joints, so I can pilot the disembodied hand around and draw the scoring functions.