C_inv.T @ C_inv -> C_inv @ C_inv.T
Ok, might be a bit of hubris, but: I think there is a typo in the IRIS paper (which is then implemented in the code and leads to a mistake... I think.)

The paper typos are in equation 5, which trickles into equation 6.
The paper (eq. 5): E = {x s.t. (x - d).T * C_inv * C_inv.T * (x - d) = 1}

Me:  E = {x s.t. (x - d).T * C_inv.T * C_inv * (x - d) = 1}

Justification:
E = {C*x + d, given |x| = 1
let u = C*x + d
x = C_inv * (u - d)
|x| = 1
|C_inv * (u - d)| = 1
x (and thus C_inv * (u - d)) is a column vector, so:
|C_inv * (u - d)| = (C_inv * (u - d)).T * (C_inv * (u - d))
= (u - d).T * C_inv.T * C_inv * (u - d))

More evidence that something is wonky:
I plotted the solution against what I came up with and there are a few tangent planes (namely the first and the last two in the linked video below) that intersect with the triangle that has the current "closest point". This seems wrong since it implies that the ellipse touches the triangle in a non-tangent manner (which in turn implies that there is a closer point than what is returned).

https://drive.google.com/file/d/1BxLxK-sO7y2qeN0eWxrEXEhZo4aL6TET/view?usp=sharing

But, I might be missing something.