Making the problem easier...
Wow! I am getting more cocky with each new commit! Wasn't sure the best way to leave feed back that I was only able to get chamfer_dist to 0.003, but that was by lower the threshold to 142 from 150.

I guess I am being lazy since there are points that satisfy the threshold dist of 150 but clearly aren't part of the mustard bottle. So I could probably write something to prune those points away... but it seems that most of the problems so far only ask for a proper implementation of whatever algorithm is presented, in this case using the heat map to generate a point cloud for the grasp.