check out and run solutions in CI
resolves #480.

one caveat is that CI always runs solutions master branch, since we
don't want to use the same git submodules machinery for the solutions
(many people who clone the repo do not have access to it, and CI needs
to check it out differently).

also had to relax the tolerance on the QP solutions to support the
open-source QP solvers.